refactor(config): enable instance validators that do not end with an underscore
The requirement of having validators end with an underscore is easily overlooked by the user of the library. Not calling validators will silently fail, which is not a good failure mode. We thus enlarge the "is a validator" test: the method name just starts with "validate_".

BREAKING CHANGE: Methods in a Config subclass that start with "validate_", do not end with an underscore, and are not validators must be renamed so that their name no longer starts with the "validate_" string.