fix(processor): made Processor.process_config private, as it related to an internal update
As we will allow configpile to process pure INI files, the `process_config` name is now ambiguous in this class.

BREAKING CHANGE: This is a breaking change as we make a public method private, but it should not break any code as there is no reason to call this method from outside the Processor class.